Moved renderDeleteDialog() into abstract class
